8146231: ppc64/gcc 4.1.2: fix build after "8143072: [JVMCI] Port JVMCI to AArch64"

+++ hotspot/src/cpu/ppc/vm/vm_version_ppc.hpp
@@ -62,7 +62,7 @@ protected:
     vcipher_m             = (1 << vcipher),
     vpmsumb_m             = (1 << vpmsumb),
     tcheck_m              = (1 << tcheck ),
-    all_features_m        = -1
+    all_features_m        = (unsigned long)-1
   };
 
   static bool _is_determine_features_test_running;
